Initial commit
Created simple server backend for a recipe app that stores
values in json file based 'data storage'. The server utilizes
'json-server' and 'lowdb' libraries in addition to JWT based
authentication library.

It exposes the custom APIs required by the recipe-app, but is just for
development purposes (i.e. should not to be used in any real system).